+++ src/components/ui/floating-navbar.tsx
@@ -8,6 +8,7 @@ import {
 } from "framer-motion";
 import { cn } from "../../lib/utils/cn";
 import { useThemeStore } from "../../context/ThemeContext";
+import { Link } from "react-router-dom";
 
 export const FloatingNav = ({
   navItems,
@@ -62,16 +63,16 @@ export const FloatingNav = ({
         )}
       >
         {navItems.map((navItem: any, idx: number) => (
-          <a
+          <Link
             key={`link=${idx}`}
-            href={navItem.link}
+            to={navItem.link}
             className={cn(
               "relative dark:text-neutral-50 items-center font-bold flex space-x-1 text-neutral-600 dark:hover:text-neutral-300 hover:text-neutral-500"
             )}
           >
             <span className="block sm:hidden">{navItem.icon}</span>
             <span className="text-sm">{navItem.name}</span>
-          </a>
+          </Link>
         ))}
         <button onClick={ToggleTheme} className="border text-sm font-medium relative border-white/[0.2] text-white px-4 py-2 rounded-full">
           <span>Login</span>
